+++ KantineAanbod.java
@@ -1,10 +1,11 @@
 import java.util.*;
+import java.math.*;
 
 public class KantineAanbod {
     // interne opslag voorraad
     private HashMap<String, ArrayList<Artikel>> aanbod;
     private HashMap<String, Integer> startVoorraad;
-    private HashMap<String, Double> prijzen;
+    private HashMap<String, BigDecimal> prijzen;
     
     /**
      * Constructor. Het eerste argument is een lijst met artikelnamen,
@@ -12,10 +13,10 @@ public class KantineAanbod {
      * is een lijst met hoeveelheden. Let op: de dimensies van de drie arrays
      * moeten wel gelijk zijn!
      */
-    public KantineAanbod(String[] artikelnaam, double[] prijs, int[] hoeveelheid) {
+    public KantineAanbod(String[] artikelnaam, BigDecimal[] prijs, int[] hoeveelheid) {
         aanbod=new HashMap<String, ArrayList<Artikel>>();
         startVoorraad=new HashMap<String, Integer>();
-        prijzen=new HashMap<String,Double>();
+        prijzen=new HashMap<String,BigDecimal>();
         for(int i=0;i<artikelnaam.length;i++) 
         {
             ArrayList<Artikel> artikelen=new ArrayList<Artikel>();
@@ -33,7 +34,7 @@ public class KantineAanbod {
     	ArrayList<Artikel> huidigeVoorraad = aanbod.get(productnaam);
     	int startHoeveelheid = startVoorraad.get(productnaam);
     	int huidigeHoeveelheid = huidigeVoorraad.size();
-    	double prijs = prijzen.get(productnaam);
+    	BigDecimal prijs = prijzen.get(productnaam);
         for(int j=huidigeHoeveelheid;j<startHoeveelheid;j++) 
         {
         	huidigeVoorraad.add(new Artikel(productnaam, prijs));
@@ -73,7 +74,7 @@ public class KantineAanbod {
     /**
      * Publieke methode om een artikel via naam van de stapel te pakken.
      * Retouneert null als artikel niet bestaat of niet op voorraad is.
-     * @param naam (van artikel)
+     * @param productnaam (van artikel)
      * @return artikel (of null)
      */
     public Artikel getArtikel(String productnaam) {
