+++ KortingskaartHouder.java
@@ -1,9 +1,11 @@
+import java.math.*;
+
 public interface KortingskaartHouder {
 
     /**
      * Methode om kortingspercentage op te vragen
      */
-    public double geefKortingsPercentage();
+    public BigDecimal geefKortingsPercentage();
 
     /**
      * Methode om op te vragen of er maximum per keer aan de korting zit
@@ -13,5 +15,5 @@ public interface KortingskaartHouder {
     /**
      * Methode om het maximum kortingsbedrag op te vragen
      */
-    public double geefMaximum();
+    public BigDecimal geefMaximum();
 }
